+++ app/queryKeys.ts
@@ -10,5 +10,4 @@ export default {
   discoverPersonas: ['discover-personas'],
   profile: (username: MaybeRef<string>) => ['profile', username],
   pricingData: ['pricing-data'],
-  stripeStatus: ['stripe-status'],
 }