+++ app/server/api/payment/stripe/pricing-data.ts
@@ -1,5 +1,5 @@
 import { PLANS } from '~/constants/payment'
-import { getPrice, getStatus } from '~/server/services/plan'
+import { getCheckoutStatus, getPrice } from '~/server/services/plan'
 import { internal } from '~/utils/nuxt'
 import { getStripe } from '~/utils/stripe'
 
@@ -13,10 +13,10 @@ export default eventHandler(async (event) => {
 
   const price = await getPrice(stripe, PLANS.allInOne)
   const user = event.context.user
-  const status = await getStatus(stripe, user)
+  const checkoutStatus = await getCheckoutStatus(stripe, user)
 
   return {
-    checkoutStatus: status?.checkoutStatus,
+    checkoutStatus,
     subscriptionStatus: user?.subscriptionStatus,
     price,
   }